fixing bad path
