+++ cmd/pentest.go
@@ -169,7 +169,7 @@ func (a *NetworkScan) InitPentestCommand() {
 	bruteForceCmd.Flags().StringSlice("passwords", []string{}, "List of passwords to use in the brute-force attack")
 	bruteForceCmd.Flags().StringSlice("username-lists", []string{}, "File paths containing lists of usernames (one per line) to use in the attack")
 	bruteForceCmd.Flags().StringSlice("password-lists", []string{}, "File paths containing lists of passwords (one per line) to use in the attack")
-	bruteForceCmd.Flags().String("wordlist-size", "", "The size of the in-built wordlist to use in the attack")
+	bruteForceCmd.Flags().String("wordlist-size", "", "The size of the in-built wordlist to use in the attack (ie. TINY)")
 	bruteForceCmd.Flags().Int("timeout", 3000, "Timeout per authentication attempt in milliseconds")
 	bruteForceCmd.Flags().Int("sleep", 3000, "Sleep duration in milliseconds between authentication attempts")
 	bruteForceCmd.Flags().Int("retries", 2, "Number of retry attempts per username/password pair")
