+++ internal/pentest/bruteforce/helpers.go
@@ -1,0 +1,104 @@
+// Package bruteforce implements password brute force functionality for various network services.
+package bruteforce
+
+import (
+	// Standard
+
+	"errors"
+	"net"
+	"strconv"
+	"strings"
+
+	// Generated
+	bruteforcefern "github.com/Method-Security/networkscan/generated/go/pentest/bruteforce"
+	// Internal
+
+	utils "github.com/Method-Security/networkscan/utils"
+)
+
+func getCredentialPairs(usernames []string, passwords []string) []bruteforcefern.CredentialPair {
+	pairs := []bruteforcefern.CredentialPair{}
+	if len(usernames) == 0 {
+		for _, password := range passwords {
+			pairs = append(pairs, bruteforcefern.CredentialPair{Password: password, Username: ""})
+		}
+		return pairs
+	}
+	for _, password := range passwords {
+		for _, username := range usernames {
+			pairs = append(pairs, bruteforcefern.CredentialPair{Password: password, Username: username})
+		}
+	}
+	return pairs
+}
+
+func getHostAndPorts(target string) (string, []int, error) {
+	var ports []int
+	host, port, err := net.SplitHostPort(target)
+	if err != nil {
+		return "", nil, err
+	}
+	intPort, err := strconv.Atoi(port)
+	if err != nil {
+		return "", nil, err
+	}
+	ports = []int{intPort}
+
+	return host, ports, nil
+}
+
+// WordlistFiles maps service types and sizes to their corresponding username and password file paths
+type WordlistFiles struct {
+	UsernameFile string
+	PasswordFile string
+}
+
+// builtInWordlistMap maps service types and wordlist sizes to their corresponding files
+var builtInWordlistMap = map[string]map[string]WordlistFiles{
+	"SSH": {
+		"TINY": {
+			UsernameFile: "conf/pentest/bruteforce/ssh/username_tiny.txt",
+			PasswordFile: "conf/pentest/bruteforce/ssh/password_tiny.txt",
+		},
+	},
+	"TELNET": {
+		"TINY": {
+			UsernameFile: "conf/pentest/bruteforce/telnet/username_tiny.txt",
+			PasswordFile: "conf/pentest/bruteforce/telnet/password_tiny.txt",
+		},
+	},
+}
+
+// GetBuiltInWordlists loads built-in username and password wordlists based on service type and size.
+// It returns slices of usernames and passwords from the appropriate wordlist files.
+func GetBuiltInWordlists(service, size string) ([]string, []string, error) {
+	// Convert service to uppercase for map lookup
+	serviceStr := strings.ToUpper(service)
+	sizeStr := strings.ToUpper(size)
+
+	// Check if service exists in the map
+	serviceMap, exists := builtInWordlistMap[serviceStr]
+	if !exists {
+		return nil, nil, errors.New("unsupported service type: " + service)
+	}
+
+	// Check if size exists for the service
+	wordlistFiles, exists := serviceMap[sizeStr]
+	if !exists {
+		return nil, nil, errors.New("unsupported wordlist size '" + size + "' for service '" + service + "'")
+	}
+
+	// Load usernames from built-in wordlist
+	usernames, err := utils.GetEntriesFromTXTFiles([]string{wordlistFiles.UsernameFile})
+	if err != nil {
+		return nil, nil, errors.New("failed to load built-in username wordlist: " + err.Error())
+	}
+
+	// Load passwords from built-in wordlist
+	passwords, err := utils.GetEntriesFromTXTFiles([]string{wordlistFiles.PasswordFile})
+	if err != nil {
+		return nil, nil, errors.New("failed to load built-in password wordlist: " + err.Error())
+	}
+
+	return usernames, passwords, nil
+}
